+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,4 +1,4 @@
-package gift.jpaRepo;
+package gift.repository;
 
 import gift.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
